More Particle fixes

+++ lib/src/frame.c
@@ -1337,7 +1337,9 @@ void enc_ack_frame(uint8_t ** pos,
                              ? DEF_ACK_DEL_EXP
                              : c->tp_out.ack_del_exp;
     const uint64_t ack_delay =
-        (uint64_t)((ev_now() - diet_timestamp(b)) * USECS_PER_SEC) >> ade;
+        (uint64_t)(
+            (uint32_t)((ev_now() - diet_timestamp(b)) * USECS_PER_SEC)) >>
+        ade;
     encv(pos, end, ack_delay);
 
     m->ack_rng_cnt = diet_cnt(&pn->recv) - 1;
